Process directives after macro-processing; fixes #1945
