+++ lib/src/model/package_graph.dart
@@ -923,7 +923,11 @@ class PackageGraph {
   }
 
   void addHtmlFragment(String name, String content) {
-    assert(!_localDocumentationBuilt);
+    // TODO(srawlins): We have to add HTML fragments after documentation is
+    // built, in order to include fragments which come from macros which
+    // were generated by a tool. I think the macro/HTML-injection system needs
+    // to be overhauled to allow for this type of looping.
+    //assert(!_localDocumentationBuilt);
     _htmlFragments[name] = content;
   }
 }
