+++ testing/test_package/bin/print_macro.dart
@@ -1,0 +1,10 @@
+// Copyright (c) 2020, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+// This is a sample "tool" used to test external tool integration into dartdoc.
+// It has no other purpose.
+
+void main() {
+  print('{@macro html-macro}');
+}
